+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "npm run test-unit; npm run test-integration;",
-    "test-unit": "NODE_ENV=test mocha --require co-mocha './server/**/*.spec.js'",
+    "test-unit": "NODE_ENV=test mocha --require co-mocha ./server/repositories/*.spec.js ./server/models/*.spec.js",
     "test-int": "NODE_ENV=test mocha --require co-mocha 'index.spec.js'",
     "test-auth": "NODE_ENV=test mocha --require co-mocha 'lib/auth.spec.js'",
     "serverTest": "cd ./server && DEBUG=loopback:*,express:* nodemon server.js",
@@ -28,6 +28,7 @@
     "jsonwebtoken": "8.1.0",
     "knex": "^0.21.5",
     "loglevel": "^1.6.8",
+    "mocha": "^8.1.3",
     "morgan": "^1.9.1",
     "nodemon": "^2.0.4",
     "pg": "^7.4.0",
@@ -41,7 +42,6 @@
     "db-migrate": "^0.10.7",
     "db-migrate-pg": "^0.4.0",
     "expect-runtime": "^0.6.0",
-    "mocha": "^2.5.3",
     "mock-knex": "^0.4.9",
     "sinon": "^1.17.7",
     "sinon-chai": "^2.8.0",
