+++ server/models/TrustModel.js
@@ -1,56 +1,38 @@
-const knex = require('../../server/database/knex');
-const expect = require("expect-runtime");
+const expect = require('expect-runtime');
 const HttpError = require("../utils/HttpError");
+const EntityRepository = require("../repositories/EntityRepository");
+const TrustRepository = require("../repositories/TrustRepository");
 
 class TrustModel{
-  async get(){
-    //const trust_relationship_instance = new trust_relationship(1);
-    const list = await knex.select()
-      .table("wallets.entity_trust");
-    return list;
+  constructor(){
+    this.trustRepository = new TrustRepository();
   }
 
-  async getById(id){
-    const list = await knex.select()
-      .table("wallets.entity_trust")
-      .where("id", id);
-    expect(list, () => new HttpError(404, 'can not find the relationship')).lengthOf.above(0);
-    expect(list, () => new HttpError(500, 'impossible, too many recodes')).lengthOf.most(1);
-    return list[0];
+  getTrustModel(){
+    return this.trustRepository;
   }
 
-  async create(trustObject){
-    await knex("wallets.entity_trust").insert(trustObject);
-    //get the inserted object, TODO there must be better way to do so.
-    const result = await knex("wallets.entity_trust").orderBy("id", "desc").limit(1);
-    return result[0];
+  async request(requestType, walletName){
+    expect(requestType, () => new HttpError(400, `The trust request type must be one of ${Object.keys(TrustRepository.ENTITY_TRUST_REQUEST_TYPE).join(',')}`))
+      .oneOf(Object.keys(TrustRepository.ENTITY_TRUST_REQUEST_TYPE));
+    expect(walletName, () => new HttpError(400, "Invalid wallet name"))
+      .match(/\S+/);
+    
+    //get wallet id
+    const entityModel = new EntityRepository();
+    const wallet = await entityModel.getEntityByWalletName((walletName));
+    const result = await this.trustRepository.create({
+      request_type: requestType,
+      target_entity_id: wallet.id,
+    });
+    return result;
   }
-
-  async update(trustObject){
-    await knex("wallets.entity_trust").update(trustObject).where("id", trustObject.id);
+  
+  async accept(trustRelationshipId){
+    const trustRelationship = await this.trustRepository.getById(trustRelationshipId);
+    trustRelationship.state = TrustRepository.ENTITY_TRUST_STATE_TYPE.trusted;
+    await this.trustRepository.update(trustRelationship);
   }
-}
-
-TrustModel.ENTITY_TRUST_TYPE = {
-  send: 'send',
-  manage: 'manage',
-  deduct: 'deduct',
-}
-
-TrustModel.ENTITY_TRUST_STATE_TYPE = {
-  requested: 'requested',
-  cancelled_by_originator: 'cancelled_by_originator',
-  canceled_by_actor: 'cancelled_by_actor',
-  trusted: 'trusted',
-}
-
-TrustModel.ENTITY_TRUST_REQUEST_TYPE = {
-  send: 'send',
-  received: 'received',
-  manage: 'manage',
-  yield: 'yield',
-  deduct: 'deduct',
-  release: 'release',
 }
 
 module.exports = TrustModel;
